Define Confirm class messages and add stream readers for Basic.Ack/Nack
The confirm.select/select-ok messages are described in https://www.rabbitmq.com/resources/specs/amqp0-9-1.extended.xml

The Basic.Ack/Nack messages are typically used by clients. However, when
publish confirmations are enabled, the broker will also send these to
the client. To this end, this commit defines the class for Basic.Nack
and adds the "fromStream" factory method so we can unmarshal these
messages when sent by the broker.