Support publish confirmations on Channel instances
This mechanism is exposed via two new channel-level methods.

The confirmPublishedMessages method enables it for a particular channel
(further invocations of the method are treated as no-ops).

The publishNotifier method allows applications to subscribe to a
stream for publish notifications. Each notification contains the
original message and message properties as well as a flag which
indicates whether the broker handled (e.g. persisted to stable storage)
or lost the message.