Ensure that calls with a noWait=true flag complete returned Futures
When an application ivokes any of the library methods with the noWait
flag set, the broker will not reply back with a message to confirm that the
request (e.g. declare a queue, exchange etc.) was successful. As a
result the returned Future values where never completed thus causing the
callers to hang.

When using the noWait flag, the broker will report errors asynchronously
via the channel or connection instance.