Allow channel.qos() invocations to set the value of the `global` param
The value of the `global` flag is interpreted differently between the AMQP 0-9-1
spec and RabbitMQ.

For RabbitMQ, if `global` is true, then `prefetchCount` is shared between
all consumers of this channel. Otherwise, `prefetchCount` is applied
separately to each new consumer of the channel.

According to the AMQP spec, if `global` is true, then `prefetchCount` is
shared between all consumers on the connection. Otherwise,
`prefetchCount` is shared between all consumers of this channel.

For more information see https://www.rabbitmq.com/consumer-prefetch.html