Implement support for heartbeats
To enable heartbeats, the client must create a new TuningSettings object
with a >= 1sec Duration value for the `heartbeatPeriod` field. The
TuningSettings must then be passed to the client's constructor via the
ConnectionSettings object.

During the initial connection handshake, the client will calculate the
effective heartbeat period (in seconds) by taking the minimum of the
heartbeat period values suggested by the client and the server.

Once a heartbeat interval has been negotiated between the client and the
server, the client will start sending heartbeat messages to the server
(approximately) twice within each heartbeat period. In addition, the
client also monitors incoming messages from the server. If no message
(heartbeat or not) has been received by the server within the heartbeat
period, the client will raise a HeartbeatFailedException.