Guard against sending heartbeats if the underlying socket is closed
If the connection to the server is lost before the library gets a chance
to detect this and cancel the heartbeat timer, there is a small window
where the client will attempt to send a heartbeat and an exception will
be thrown.

This commit attempts to guard against this by checking the channel and
socket status before attempting to transmit the heartbeat message as
well as by wrapping the send request in a try/catch block and
suppressing any thrown exceptions.