Introduce maxMissedHeartbeats tuning parameter
This commit implements a suggestion from https://github.com/achilleasa/dart_amqp/issues/106#issuecomment-1846280852.

The client-side heartbeat implementation is modified via the
introduction of a new tuning parameter called `maxMissedHeartbeats`
(default: 3).

If heartbeats are enabled and the client sends `maxMissedHeartbeats`
consecutive heartbeat messages at the interval negotiated with the
broker without receiving any message (heartbeat or regular traffic)
back, the client assumes that the server is not available and will raise
a HeartbeatFailedException.