Do not close error stream unless client is explicitly closed via close()
This change addresses one of the issues described in
https://github.com/achilleasa/dart_amqp/issues/107.

Prior to this change, the error broadcast stream was always closed when
the connection was closed either by the client or the server. This
prevented clients that implemented a reconnection policy from being able
to register error listeners.

This commit modifies the client behavior so that the error broadcast
channel is only closed if the client invokes the close() method on a
client instance.